fix(storage): use custom precache strategy for initialization
As Models for Storage are entangled (eq. has dependencies between each others)
the order of processing is important as things might not exists when needed.
EventEmitters does not await competition of processing an event so this is not guaranteed.
Using the precache function guarantees order of processing so it is used instead of the watcher's past-event processing strategy.