fix(cli): oclif release command path modification
Oclif has specified path where to look for commands in the package.json.
In development environment it should look into the /src/ path, but for
released packages it has to be into compiled folder /lib/. There is some
magic in Oclif that should make this work, but it does not work properly
so it needs to be changed manually.

Hacky fix is replacing this value before releasing to NPM and then
returning this value after release if finished.